backend/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose
pydantic
sqlalchemy
python-dotenv
email-validator